Fix u16 overflow in 2-bit and 4-bit bitmap decoders
The index calculations (y * scan_width + x) were done in u16 arithmetic,
causing overflow panics for bitmaps where the product exceeds 65535.
Cast to usize before multiplying, matching the existing fix in the 1-bit decoder.

+++ vm-rust/src/player/bitmap/bitmap.rs
@@ -293,7 +293,7 @@ fn decode_bitmap_2bit(
     let mut result_bmp = vec![0; width as usize * height as usize];
     for y in 0..scan_height {
         for x in 0..scan_width {
-            let compressed_index = (y * scan_width + x) as usize;
+            let compressed_index = y as usize * scan_width as usize + x as usize;
             if compressed_index >= decoded_data.len() {
                 return Err(format!(
                     "decode_bitmap_2bit: compressed_index {} >= decoded_data.len() {}",
@@ -302,7 +302,7 @@ fn decode_bitmap_2bit(
                 ));
             }
             if x < width {
-                let pixel_index = (y * width + x) as usize;
+                let pixel_index = y as usize * width as usize + x as usize;
                 let pixel = decoded_data[compressed_index];
                 result_bmp[pixel_index] = pixel;
             }
@@ -349,7 +349,7 @@ fn decode_bitmap_4bit(
 
     for y in 0..height {
         for x in 0..width {
-            let scan_index = (y * scan_width + x) as usize;
+            let scan_index = y as usize * scan_width as usize + x as usize;
 
             if scan_index >= decoded_data.len() {
                 return Err(format!(
@@ -360,7 +360,7 @@ fn decode_bitmap_4bit(
             }
 
             let pixel = decoded_data[scan_index];
-            let pixel_index = (y * width + x) as usize;
+            let pixel_index = y as usize * width as usize + x as usize;
 
             // Store as 8-bit indexed (values 0-15)
             result_bmp[pixel_index] = pixel;
